# Reduce Any stack usage with map + list of lookups