Alternative to the other Any stack optimizations,  building on top of the tests in https://github.com/uber-go/zap/pull/1303.

This is a different approach to the Any function using a prebuilt a map by type to determine which function to call. The map only works for concrete types, so we also need a list for interfaces, ordered in precedence order (since multiple interfaces may match a single type).

I'm using the output of `go build -gcflags -S` to determine the stack frame size, which allows us to compare the approaches.
```
# master
go.uber.org/zap.Any STEXT size=16672 args=0x20 locals=0x1308 funcid=0x0 align=0x0
        0x0000 00000 (/Users/prashant/dev/3p/zap/field.go:420)  TEXT    go.uber.org/zap.Any(SB), ABIInternal, $4880-32
```
In the above, locals is 0x1308 = 4872, which is 8 bytes less than the frame size on the next line ($4880).

Using this, we can compare the different approaches directly,
```
PR #1301:
go.uber.org/zap.Any STEXT size=5072 args=0x20 locals=0x3e8 funcid=0x0 align=0x0
        0x0000 00000 (/Users/prashant/dev/3p/zap/field.go:421)  TEXT    go.uber.org/zap.Any(SB), ABIInternal, $1008-32

PR #1302:
go.uber.org/zap.Any STEXT size=4048 args=0x20 locals=0x148 funcid=0x0 align=0x0
        0x0000 00000 (/Users/prashant/dev/3p/zap/field.go:422)  TEXT    go.uber.org/zap.Any(SB), ABIInternal, $336-32

PR #1303
go.uber.org/zap.Any STEXT size=6320 args=0x20 locals=0x648 funcid=0x0 align=0x0
        0x0000 00000 (/Users/prashant/dev/3p/zap/field.go:420)  TEXT    go.uber.org/zap.Any(SB), ABIInternal, $1616-32

PR #1304:
go.uber.org/zap.Any STEXT size=3664 args=0x20 locals=0x128 funcid=0x0 align=0x0
        0x0000 00000 (/Users/prashant/dev/3p/zap/field.go:428)  TEXT    go.uber.org/zap.Any(SB), ABIInternal, $304-32

PR #1305:
go.uber.org/zap.Any STEXT size=7120 args=0x20 locals=0xb8 funcid=0x0 align=0x0
        0x0000 00000 (/Users/prashant/dev/3p/zap/field.go:420)  TEXT    go.uber.org/zap.Any(SB), ABIInternal, $192-32

Branch pawel/any-int5
go.uber.org/zap.Any STEXT size=7216 args=0x20 locals=0xc8 funcid=0x0 align=0x0
        0x0000 00000 (/Users/prashant/dev/3p/zap/field.go:420)  TEXT    go.uber.org/zap.Any(SB), ABIInternal, $208-32

This PR:
go.uber.org/zap.Any STEXT size=448 args=0x20 locals=0xf8 funcid=0x0 align=0x0
        0x0000 00000 (/Users/prashant/dev/3p/zap/field.go:506)  TEXT    go.uber.org/zap.Any(SB), ABIInternal, $256-32
```

Benchmarks for this version show `Any` and `String` perform similarly:
```
BenchmarkAny/str-in-go            782828              1422 ns/op              89 B/op          2 allocs/op
BenchmarkAny/any-in-go            723226              1446 ns/op              88 B/op          2 allocs/op
BenchmarkAny/str-in-go-with-stack                 887086              1436 ns/op              88 B/op          2 allocs/op
BenchmarkAny/any-in-go-with-stack                 757072              1455 ns/op              88 B/op          2 allocs/op
```