@rabbbit 2023-07-29
An actual question, not a hidden message - I really don't know.

How do these locals+stack sizes work wtr to nested functions? I imagine they wouldn't show up as stack size, possibly not as locals too? Is there an easy way to measure the whole thing?

What does the `size` mean here? You're clearly winning in this one :)