@prashantv 2023-07-30
> How do these locals+stack sizes work wtr to nested functons? I imagine they wouldn't show up as stack size, possibly not as locals too? Is there an easy way to measure the whole thing?

That's right, these only measure the cost of the specific function, not the nested calls as far as I know.

> What does the `size` mean here? You're clearly winning in this one :)

Size is the amount of code directly in the function. With the switch, all the cases are in code, there's a lot of tests + branches to get to the right case. With this approach, we've encoded the cases into data (hashmap), and the lookup logic is also through a separate call, so it makes the size smaller, but it doesn't really mean a huge amount (less cache impact, but we're also calling other functions which could counteract that).

I think the benchmarks you added are the best way to measure the impact, wdyt about landing those as a separate PR since it's useful to have the same benchmarks to compare the different approaches? We may want to add a benchmark that does Any with many types to measure the impact of the different branches as well.