@rabbbit 2023-07-30
> I think the benchmarks you added are the best way to measure the impact, wdyt about landing those as a separate PR since it's useful to have the same benchmarks to compare the different approaches? We may want to add a benchmark that does Any with many types to measure the impact of the different branches as well.

SGTM, can put it up. I opted for `string` for now since more complicated things were allocating more so the impact wasn't as obvious. Maybe we're pairing too negative picture then.