@prashantv 2023-08-09
Closing since this was primarily intended to improve the stack usage, which #1310 brings in a more minimal way.

This may be worth coming back to for maintainability and possibly other performance reasons, but that should be considered separately.